+++ package.json
@@ -10,7 +10,8 @@
         "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm run build:tailwind && npm run serve",
         "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
         "lint:fix": "eslint --ext .js,.vue --ignore-path .gitignore . --fix",
-        "prepare": "husky install"
+        "prepare": "husky install",
+        "test": ""
     },
     "dependencies": {
         "@fortawesome/fontawesome-free": "5.14.0",
