+++ src/views/admin/AdminDashboard.vue
@@ -1,6 +1,10 @@
 <template>
   <div>
-    <ul id="tabs" class="inline-flex w-full px-1 pt-2 " v-if="canView === 'beta'">
+    <ul
+      id="tabs"
+      class="inline-flex w-full px-1 pt-2"
+      v-if="canView === 'beta'"
+    >
       <li
         class="px-4 py-2 -mb-px font-semibold text-gray-800 border-b-2 border-blue-400 rounded-t opacity-50"
       >
@@ -12,7 +16,7 @@
     </ul>
     <div id="tab-contents">
       <div id="first">
-        <div class=" flex flex-row">
+        <div class="flex flex-row">
           <div
             class="relative flex flex-col min-w-0 break-words bg-white w-full mb-6 shadow-lg rounded pr-2"
           >
@@ -145,7 +149,7 @@
             </div>
           </div>
         </div>
-        <div class=" flex flex-row">
+        <div class="flex flex-row">
           <div
             class="relative flex flex-col min-w-0 break-words bg-white w-full mb-6 shadow-lg rounded pr-2"
           >
@@ -202,7 +206,7 @@
                   </tr>
                 </thead>
                 <tbody>
-                  <tr v-for="(user, index) in agents">
+                  <tr v-for="(user, index) in agents" :key="user.id">
                     <th
                       class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-no-wrap p-4 text-left"
                     >
@@ -242,96 +246,115 @@
         </div>
       </div>
       <div id="second" v-if="canView === 'beta'">
-        <un-converted/>
+        <un-converted />
       </div>
     </div>
   </div>
 </template>
 
-
 <script>
-import ShowRoomStats from "../../components/ShowRoomStats.vue";
-import { get } from "../../utilities/api.js";
-import showroomData from "../../utilities/chartData.js";
-import queryParam from "../../utilities/queryParam";
-import UnConverted from "../../components/UnConverted.vue";
+  import ShowRoomStats from "../../components/ShowRoomStats.vue";
+  import { get } from "../../utilities/api.js";
+  import showroomData from "../../utilities/chartData.js";
+  import queryParam from "../../utilities/queryParam";
+  import UnConverted from "../../components/UnConverted.vue";
 
-export default {
-  name: "AdminDashboard",
-  components: {
-    ShowRoomStats,
-    UnConverted
-  },
-  data() {
-    return {
-      apiUrls: {
-        feedbacks: `/api/feedbacks`,
-        getDSAs: `/api/get-users?role=18&stats=true`
+  export default {
+    name: "AdminDashboard",
+    components: {
+      ShowRoomStats,
+      UnConverted,
+    },
+    data() {
+      return {
+        apiUrls: {
+          feedbacks: `/api/feedbacks`,
+          getDSAs: `/api/get-users?role=18&stats=true`,
+        },
+        graphData: showroomData,
+        pageParams: {},
+        agents: [],
+        searchQuery: {},
+        canView: "",
+      };
+    },
+
+    mounted() {
+      this.getAgents();
+      this.canView = localStorage.getItem("flag");
+    },
+
+    methods: {
+      async getAgents() {
+        this.$LIPS(true);
+        try {
+          const query = {
+            ...this.searchQuery,
+            page: this.pageParams.page,
+            limit: this.pageParams.limit,
+          };
+          const agents = await get(this.apiUrls.getDSAs + queryParam(query));
+          this.agents = agents.data.data.data;
+          let {
+            current_page,
+            first_page_url,
+            from,
+            last_page,
+            last_page_url,
+            per_page,
+            next_page_url,
+            to,
+            total,
+            prev_page_url,
+          } = agents.data.data;
+          this.pageParams = Object.assign({}, this.pageParams, {
+            current_page,
+            first_page_url,
+            from,
+            last_page,
+            last_page_url,
+            per_page,
+            next_page_url,
+            to,
+            total,
+            prev_page_url,
+          });
+        } catch (err) {
+          this.$displayErrorMessage(err);
+        } finally {
+          this.$LIPS(false);
+        }
       },
-      graphData: showroomData,
-      pageParams: {},
-      agents: [],
-      searchQuery: {},
-      canView: ""
-    };
-  },
 
-  mounted() {
-    this.getAgents();
-    this.canView = localStorage.getItem("flag");
-  },
+      next(firstPage = null) {
+        if (this.pageParams.next_page_url) {
+          this.pageParams.page = firstPage
+            ? firstPage
+            : parseInt(this.pageParams.page) + 1;
+          this.$router.push({
+            path: this.$route.path,
+            query: {
+              ...this.$route.query,
+              page: this.pageParam.page,
+            },
+          });
+          this.getAgents();
+        }
+      },
 
-  methods: {
-    async getAgents() {
-      console.log("work");
-      this.$LIPS(true);
-      try {
-        const query = {
-          ...this.searchQuery,
-          page: this.pageParams.page,
-          limit: this.pageParams.limit
-        };
-        const agents = await get(this.apiUrls.getDSAs + queryParam(query));
-        console.log("work");
-        this.agents = agents.data.data.data;
-        let {
-          current_page,
-          first_page_url,
-          from,
-          last_page,
-          last_page_url,
-          data,
-          per_page,
-          next_page_url,
-          to,
-          total,
-          prev_page_url
-        } = agents.data.data;
-        this.pageParams = Object.assign({}, this.pageParams, {
-          current_page,
-          first_page_url,
-          from,
-          last_page,
-          last_page_url,
-          per_page,
-          next_page_url,
-          to,
-          total,
-          prev_page_url
-        });
-      } catch (err) {
-        this.$displayErrorMessage(err);
-      } finally {
-        this.$LIPS(false);
-      }
-    }
-  }
-};
+      previous(lastPage = null) {
+        if (this.param.prev_page_url) {
+          this.param.page = lastPage ? lastPage : parseInt(this.param.page) - 1;
+          this.$router.push({
+            path: this.$route.path,
+            query: {
+              ...this.$route.query,
+              page: this.param.page,
+            },
+          });
+          this.getAgents();
+        }
+      },
+    },
+  };
 </script>
-
-
-
- 
-
-
-
